Fix: fmi2ValueReferenceFormat/fmi3ValueReferenceFormat must use Int, not Int64
Same bug class as the cRef fix: these Union type aliases hardcoded
Int64 to accept a plain-integer value-reference shorthand (e.g.
getValue(sol, 1)). Integer literals default to Julia's native Int,
which is Int32 on 32-bit -- so on 32-bit Julia, passing a bare integer
or Vector{Int} didn't match the Union at all, and dispatch silently
fell through to an unrelated method missing the `isIndex` keyword,
producing a confusing MethodError instead of doing the intended thing.

Using Int (word-size-native) matches integer literals on both
architectures.

Found via FMIBase's own test suite (struct_solution.jl) failing on
Julia 1.12.7 x86 after the cRef fix was applied -- 2 errors there,
now resolved. Full non-GR-dependent suite: 253 passed, 1 broken
(pre-existing, deliberate skip for missing 32-bit demo FMU binary,
unrelated to this change), 0 errors.

+++ src/base.jl
@@ -30,8 +30,8 @@ const fmi2ValueReferenceFormat = Union{
     AbstractArray{String,1},
     fmi2ValueReference,
     AbstractArray{fmi2ValueReference,1},
-    Int64,
-    AbstractArray{Int64,1},
+    Int,
+    AbstractArray{Int,1},
     Symbol,
 }
 const fmi3ValueReferenceFormat = Union{
@@ -40,8 +40,8 @@ const fmi3ValueReferenceFormat = Union{
     AbstractArray{String,1},
     fmi3ValueReference,
     AbstractArray{fmi3ValueReference,1},
-    Int64,
-    AbstractArray{Int64,1},
+    Int,
+    AbstractArray{Int,1},
     Symbol,
 }
 const fmiValueReferenceFormat = Union{fmi2ValueReferenceFormat,fmi3ValueReferenceFormat}
